[Feat] 자동저장 기능구현

+++ Assets/Project/KimEusung/Scripts/Save/GameData.cs
@@ -6,7 +6,7 @@ using System.IO;
 [System.Serializable]
 public class GameData
 {
-    public int playerHp;             // 플레이어의 생명갯수
+    public float playerHp;             // 플레이어의 생명갯수
     public Vector3 playerPosition; // 플레이어의 위치를 저장함 (상황에맞게 사용 Vector3, Vector2)
     //public Vector2 playerPosition;
     public List<string> items;       // 플레이어 아이템 목록 (나중에 구현필요)
